Modified 1-print_binary.c

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -15,7 +15,7 @@
 void print_binary(unsigned long int n)
 {
 	int shft;
-	unsigned int tmp;
+	unsigned long int tmp;
 
 	if (n == 0)
 	{
